# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Observ.ing API",
    "description": "HTTP API behind [observ.ing](https://observ.ing), a biodiversity observation app built on the AT Protocol.\n\nReads come from the appview's index of the network. Writes go to the signed-in user's PDS as AT Protocol records, then reach the index via the firehose, so a newly written record can take a moment to appear in reads.\n\nAuthentication is the `session_did` cookie set by the OAuth flow (`/oauth/login`). Endpoints that accept it optionally use it to fill in viewer-specific fields such as `viewerHasLiked`.",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/interactions": {
      "post": {
        "tags": [
          "interactions"
        ],
        "summary": "Record a species interaction.",
        "description": "Writes an interaction record (e.g. predation, pollination) between two\nsubjects to the caller's PDS.",
        "operationId": "create_interaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInteractionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The record was written to the caller's PDS",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "A field failed validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/interactions/occurrence/{uri}": {
      "get": {
        "tags": [
          "interactions"
        ],
        "summary": "List species interactions involving an occurrence.",
        "operationId": "list_interactions",
        "parameters": [
          {
            "name": "uri",
            "in": "path",
            "description": "AT URI of the occurrence (`at://...`), percent-encoded",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Interactions in which the occurrence is either subject",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionListResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateInteractionRequest": {
        "type": "object",
        "required": [
          "subjectA",
          "subjectB",
          "interactionType"
        ],
        "properties": {
          "subjectA": {
            "$ref": "#/components/schemas/InteractionSubjectRequest"
          },
          "subjectB": {
            "$ref": "#/components/schemas/InteractionSubjectRequest"
          },
          "interactionType": {
            "type": "string"
          },
          "direction": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/InteractionDirection",
                "description": "Defaults to `AtoB`."
              },
              {
                "type": "null"
              }
            ]
          },
          "comment": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "EnrichedInteraction": {
        "allOf": [
          {
            "$ref": "#/components/schemas/InteractionRow"
          },
          {
            "type": "object",
            "required": [
              "creator"
            ],
            "properties": {
              "creator": {
                "$ref": "#/components/schemas/Profile"
              }
            }
          }
        ],
        "description": "Enriched interaction with profile info"
      },
      "ErrorResponse": {
        "type": "object",
        "description": "JSON body of every error response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "InteractionDirection": {
        "type": "string",
        "enum": [
          "AtoB",
          "BtoA",
          "bidirectional"
        ]
      },
      "InteractionListResponse": {
        "type": "object",
        "required": [
          "interactions"
        ],
        "properties": {
          "interactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EnrichedInteraction"
            }
          }
        }
      },
      "InteractionRow": {
        "type": "object",
        "description": "Interaction row returned from SELECT queries",
        "required": [
          "uri",
          "cid",
          "did",
          "interaction_type",
          "direction",
          "created_at"
        ],
        "properties": {
          "uri": {
            "type": "string"
          },
          "cid": {
            "type": "string"
          },
          "did": {
            "type": "string"
          },
          "subject_a_occurrence_uri": {
            "type": [
              "string",
              "null"
            ]
          },
          "subject_a_taxon_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "subject_a_kingdom": {
            "type": [
              "string",
              "null"
            ]
          },
          "subject_b_occurrence_uri": {
            "type": [
              "string",
              "null"
            ]
          },
          "subject_b_taxon_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "subject_b_kingdom": {
            "type": [
              "string",
              "null"
            ]
          },
          "interaction_type": {
            "type": "string"
          },
          "direction": {
            "$ref": "#/components/schemas/InteractionDirection"
          },
          "comment": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "indexed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "InteractionSubjectRequest": {
        "type": "object",
        "description": "One side of an interaction: an occurrence, a taxon, or both.",
        "properties": {
          "occurrenceUri": {
            "type": [
              "string",
              "null"
            ]
          },
          "occurrenceCid": {
            "type": [
              "string",
              "null"
            ]
          },
          "scientificName": {
            "type": [
              "string",
              "null"
            ]
          },
          "kingdom": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "Profile": {
        "type": "object",
        "required": [
          "did"
        ],
        "properties": {
          "did": {
            "type": "string"
          },
          "handle": {
            "type": [
              "string",
              "null"
            ]
          },
          "displayName": {
            "type": [
              "string",
              "null"
            ]
          },
          "avatar": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "RecordCreatedResponse": {
        "type": "object",
        "description": "Response returned when an AT Protocol record is created.",
        "required": [
          "success",
          "uri",
          "cid"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "uri": {
            "type": "string"
          },
          "cid": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_did",
        "description": "Set by the OAuth callback after signing in via `/oauth/login`."
      }
    }
  },
...
}